- Follow Maven convention. Move source files from src to src/main/java
